refactor: refactor primitive mesh
