--- packages/core/src/geometry-shape/ScreenQuadGeometry.ts
@@ -1,30 +1,0 @@
-import { Engine } from "../Engine";
-import { MeshTopology } from "../graphic/enums/MeshTopology";
-import { VertexElementFormat } from "../graphic/enums/VertexElementFormat";
-import { VertexElement } from "../graphic/VertexElement";
-import { ShapeGeometry } from "./ShapeGeometry";
-
-/**
- * A rectangle covering the entire screen.
- * @private
- */
-export class ScreenQuadGeometry extends ShapeGeometry {
-  constructor(engine: Engine) {
-    super(engine);
-    const vertices: Float32Array = new Float32Array([-1, -1, 0, 0, 0, 1, -1, 0, 1, 0, 1, 1, 0, 1, 1, -1, 1, 0, 0, 1]);
-
-    const indices: Uint16Array = new Uint16Array([0, 1, 2, 3]);
-
-    this._initialize(engine, vertices, indices);
-    this.subMesh.topology = MeshTopology.TriangleFan;
-  }
-
-  _initialize(engine: Engine, vertices: Float32Array, indices: Uint16Array) {
-    const vertexStride = 20;
-    const vertexElements = [
-      new VertexElement("POSITION", 0, VertexElementFormat.Vector3, 0),
-      new VertexElement("TEXCOORD_0", 12, VertexElementFormat.Vector2, 0)
-    ];
-    this._initBuffer(engine, vertices, indices, vertexStride, vertexElements);
-  }
-}
