--- packages/core/src/geometry-shape/index.ts
@@ -1,5 +1,0 @@
-export { CylinderGeometry } from "./CylinderGeometry";
-export { ScreenQuadGeometry } from "./ScreenQuadGeometry";
-export { ShapeGeometry } from "./ShapeGeometry";
-export { PrimitiveMesh } from "./PrimitiveMesh";
-
