+++ packages/core/src/mesh/index.ts
@@ -1,0 +1,6 @@
+export { ModelMesh } from "./ModelMesh";
+export { Skin } from "./Skin";
+export { MeshRenderer } from "./MeshRenderer";
+export { SkinnedMeshRenderer } from "./SkinnedMeshRenderer";
+export { LODGroup } from "./LODGroup";
+export { PrimitiveMesh } from "./PrimitiveMesh";
